Add build script option to skip autotools
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/dev@6425 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ cutil_build.sh
@@ -13,13 +13,16 @@ EXTRA_ARGS=
 if [ ! -z "$EXM_OPT_BUILD" ]; then
     EXTRA_ARGS+="--enable-fast"
 fi
-rm -rf ./config.status ./autom4te.cache
+
+if [ -z "$SKIP_AUTOTOOLS" ]; then
+  rm -rf ./config.status ./autom4te.cache
+  ./setup.sh
+fi
 
 if [ ! -z "$EXM_DEBUG_BUILD" ]; then
    export CFLAGS="-g -O0" 
 fi
 
-./setup.sh
 ./configure --enable-shared --prefix=${C_UTILS_INST} ${EXTRA_ARGS} CC=gcc
 make -j ${MAKE_PARALLELISM}
 make install
